+++ 3pt5_initsRefs/References/references_mess.cpp
@@ -35,7 +35,7 @@ void ref_var_example(){
 
 void print_vector(vector<int> vec){
 	cout << "{ ";
-	for(int i = 0; i < vec.size() - 1; i++){
+	for(size_t i = 0; i < vec.size() - 1; i++){
 		cout << vec[i] << ", ";
 	}
 
@@ -44,7 +44,7 @@ void print_vector(vector<int> vec){
 
 void print_vector(vector<std::pair<int, int>> vec){
 	cout << "{ ";
-	for(int i = 0; i < vec.size() - 1; i++){
+	for(size_t i = 0; i < vec.size() - 1; i++){
 		cout << "{" << vec[i].first << ", " << vec[i].second << "}, ";
 	}
